Move wip_units_resource_loader.* code into unitsrouter.*
